requirements.txt
discord.py>=2.2.0
motor==3.1.1